+++ test_more/general/div.cpp
@@ -141,7 +141,7 @@ bool DivTestTwo(void)
 	double value = 6. / 4.;
 	for(j = 1; j < p; j++)
 	{
-		jfac *= j;
+		jfac *= double(j);
 		w     = f.Forward(j, v);
 		ok &= NearEqual(w[0], value/jfac, eps99, eps99); // d^jz/du^j
 		v[0]  = 0.;
@@ -157,7 +157,7 @@ bool DivTestTwo(void)
 	for(j = 0; j < p; j++)
 	{
 		ok &= NearEqual(r[j], value/jfac, eps99, eps99); // d^jz/du^j
-		jfac *= (j + 1);
+		jfac *= double(j + 1);
 		value = 0.;
 	}
