+++ test_more/general/sin.cpp
@@ -58,7 +58,7 @@ bool Sin(void)
 			value = -cos_u;
 		else	value = sin_u;
 
-		jfac *= j;
+		jfac *= double(j);
 		ok &= NearEqual(w[0], value/jfac, eps99, eps99); // d^jz/du^j
 		v[0]  = 0.;
 	}
@@ -81,7 +81,7 @@ bool Sin(void)
 
 		ok &= NearEqual(r[j], value/jfac, eps99, eps99); // d^jz/du^j
 
-		jfac *= (j + 1);
+		jfac *= double(j + 1);
 	}
 
 	return ok;
