+++ test_more/general/sinh.cpp
@@ -54,7 +54,7 @@ bool Sinh(void)
 			value = cos_u;
 		else	value = sin_u;
 
-		jfac *= j;
+		jfac *= double(j);
 		ok &= NearEqual(w[0], value/jfac, eps99, eps99); // d^jz/du^j
 		v[0]  = 0.;
 	}
@@ -73,7 +73,7 @@ bool Sinh(void)
 
 		ok &= NearEqual(r[j], value/jfac, eps99, eps99); // d^jz/du^j
 
-		jfac *= (j + 1);
+		jfac *= double(j + 1);
 	}
 
 	return ok;
